+++ R/scale_process.R
@@ -13,7 +13,7 @@
 #' @author Insang Song
 #' @examples
 #' err <- simpleError("No input.")
-#' par_fallback(err, extract_at, debug = TRUE, 1)
+#' par_fallback(err, extract_at, inputid = 1)
 #' @export
 par_fallback <-
   function(
@@ -40,8 +40,8 @@ par_fallback <-
 #' @family Parallelization
 #' @description
 #' [future::multicore], [future::multisession], [future::cluster]
-#' with [doParallel::registerDoParallel] will parallelize the work
-#' in each grid. For details of the terminology in \code{future} package,
+#' will parallelize the work in each grid.
+#' For details of the terminology in \code{future} package,
 #' refer to \link[future]{plan}. This function assumes that
 #' users have one raster file and a sizable and spatially distributed
 #' target locations. Each thread will process
@@ -258,12 +258,11 @@ par_grid <-
 #'  which divides the entire study region into multiple subregions.
 #'  It is oftentimes reflected in an area code system
 #'  (e.g., FIPS for US Census geographies, HUC-4, -6, -8, etc.).
-#' [future::multicore], [future::multisession], [future::cluster]
-#' with [doParallel::registerDoParallel] will parallelize the work
-#' in each grid. For details of the terminology in \code{future} package,
+#'  [future::multicore], [future::multisession], [future::cluster]
+#'  will parallelize the work by splitting lower level features into
+#'  several higher level feature group.
+#'  For details of the terminology in \code{future} package,
 #'  refer to \link[future]{plan}.
-#'  This function assumes that users have one raster file and
-#'  a sizable and spatially distributed target locations.
 #'  Each thread will process the number of lower level features
 #'  in each higher level feature. Please be advised that
 #'  accessing the same file simultaneously with